chore: Convert tabs to index based stories
