+++ packages/react-components/react-tabs/src/stories/TabList/TabListDisabled.stories.tsx
@@ -1,7 +1,7 @@
 import { CalendarMonthRegular } from '@fluentui/react-icons';
 import { makeStyles, shorthands } from '@griffel/react';
 import * as React from 'react';
-import { Tab, TabList } from '../index';
+import { Tab, TabList } from '@fluentui/react-tabs';
 
 const useStyles = makeStyles({
   root: {
