+++ packages/react-components/react-tabs/src/stories/TabList/TabListIconOnly.stories.tsx
@@ -1,6 +1,6 @@
 import { makeStyles, shorthands } from '@griffel/react';
 import * as React from 'react';
-import { Tab, TabList } from '../index';
+import { Tab, TabList } from '@fluentui/react-tabs';
 import { CalendarMonthRegular } from '@fluentui/react-icons';
 
 const useStyles = makeStyles({
