+++ packages/react-components/react-tabs/src/stories/TabList/TabListSizeSmall.stories.tsx
@@ -1,7 +1,7 @@
+import { CalendarMonthRegular } from '@fluentui/react-icons';
 import { makeStyles, shorthands } from '@griffel/react';
 import * as React from 'react';
-import { Tab, TabList } from '../index';
-import { CalendarMonthRegular } from '@fluentui/react-icons';
+import { Tab, TabList } from '@fluentui/react-tabs';
 
 const useStyles = makeStyles({
   root: {
@@ -14,42 +14,38 @@ const useStyles = makeStyles({
   },
 });
 
-export const WithIcon = () => {
+export const SizeSmall = () => {
   const styles = useStyles();
 
   const renderTabs = () => {
     return (
       <>
-        <Tab icon={<CalendarMonthRegular />} value="tab1">
-          First Tab
-        </Tab>
+        <Tab value="tab1">First Tab</Tab>
         <Tab icon={<CalendarMonthRegular />} value="tab2">
           Second Tab
         </Tab>
-        <Tab icon={<CalendarMonthRegular />} value="tab3">
-          Third Tab
-        </Tab>
-        <Tab icon={<CalendarMonthRegular />} value="tab4">
-          Fourth Tab
-        </Tab>
+        <Tab value="tab3">Third Tab</Tab>
+        <Tab value="tab4">Fourth Tab</Tab>
       </>
     );
   };
 
   return (
     <div className={styles.root}>
-      <TabList defaultSelectedValue="tab2">{renderTabs()}</TabList>
-      <TabList defaultSelectedValue="tab2" vertical>
+      <TabList defaultSelectedValue="tab2" size="small">
+        {renderTabs()}
+      </TabList>
+      <TabList defaultSelectedValue="tab2" size="small" vertical>
         {renderTabs()}
       </TabList>
     </div>
   );
 };
 
-WithIcon.parameters = {
+SizeSmall.parameters = {
   docs: {
     description: {
-      story: 'A tab has an `icon` slot to display an icon before the tab content.',
+      story: 'A tab list can have `small` tabs. The default size is `medium`.',
     },
   },
 };
