+++ packages/react-components/react-tabs/src/stories/TabList/TabListWithIcon.stories.tsx
@@ -1,7 +1,7 @@
-import { CalendarMonthRegular } from '@fluentui/react-icons';
 import { makeStyles, shorthands } from '@griffel/react';
 import * as React from 'react';
-import { Tab, TabList } from '../index';
+import { Tab, TabList } from '@fluentui/react-tabs';
+import { CalendarMonthRegular } from '@fluentui/react-icons';
 
 const useStyles = makeStyles({
   root: {
@@ -14,38 +14,42 @@ const useStyles = makeStyles({
   },
 });
 
-export const SizeMedium = () => {
+export const WithIcon = () => {
   const styles = useStyles();
 
   const renderTabs = () => {
     return (
       <>
-        <Tab value="tab1">First Tab</Tab>
+        <Tab icon={<CalendarMonthRegular />} value="tab1">
+          First Tab
+        </Tab>
         <Tab icon={<CalendarMonthRegular />} value="tab2">
           Second Tab
         </Tab>
-        <Tab value="tab3">Third Tab</Tab>
-        <Tab value="tab4">Fourth Tab</Tab>
+        <Tab icon={<CalendarMonthRegular />} value="tab3">
+          Third Tab
+        </Tab>
+        <Tab icon={<CalendarMonthRegular />} value="tab4">
+          Fourth Tab
+        </Tab>
       </>
     );
   };
 
   return (
     <div className={styles.root}>
-      <TabList defaultSelectedValue="tab2" size="medium">
-        {renderTabs()}
-      </TabList>
-      <TabList defaultSelectedValue="tab2" size="medium" vertical>
+      <TabList defaultSelectedValue="tab2">{renderTabs()}</TabList>
+      <TabList defaultSelectedValue="tab2" vertical>
         {renderTabs()}
       </TabList>
     </div>
   );
 };
 
-SizeMedium.parameters = {
+WithIcon.parameters = {
   docs: {
     description: {
-      story: 'A tab list can have `medium` tabs. The default size is `medium`.',
+      story: 'A tab has an `icon` slot to display an icon before the tab content.',
     },
   },
 };
